--- src/main/java/com/mn/matdi/dto/EmailVerificationNumberDto.java
@@ -1,15 +1,0 @@
-package com.mn.matdi.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class EmailVerificationNumberDto {
-    private String email;
-    private String vrfNo;
-}
