+++ src/main/java/com/mn/matdi/dto/user/UserSignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.mn.matdi.dto;
+package com.mn.matdi.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
