+++ src/main/java/com/mn/matdi/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.mn.matdi.mapper;
 
-import com.mn.matdi.dto.UserSignUpRequestDto;
+import com.mn.matdi.dto.user.UserSignUpRequestDto;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
