+++ src/main/resources/mapper/user.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.mn.matdi.mapper.UserMapper">
-    <insert id="insertUser" parameterType="com.mn.matdi.dto.UserSignUpRequestDto">
+    <insert id="insertUser" parameterType="com.mn.matdi.dto.user.UserSignUpRequestDto">
         INSERT INTO user (
                 email,
                 user_pwd,
